Review Server Network and Output dashboard
Recursively travels through the server network and at each server, evaluates if the server has been cracked. Outputs security and money for the server if so, and the requirements to hack the server otherwise. If it finds a hackable server, it will ask if it should run autoHack.ns, and if so, it will automatically hack.